tagging 1.4.3 version

* tag '1.4.3' of https://github.com/exfatprogs/exfatprogs:
  exfatprogs: release 1.4.3 version
  libexfat: avoid overflow in file size calculation
  fsck: optimize unused dentry scan
  fsck: scan unused dentries in larger reads
  libexfat: avoid overflow in bitmap size calculation
  README: direct feedback to the exfat mailing list
  libexfat: fix invalid bitwise aligning in exfat_de_iter_revert()
  libexfat: increase bitmap_t to wordsize
  lib: add internal library tests
  libexfat: fix logic error in exfat_count_used_clusters()
  mkfs: fix GUID generation
  mkfs: tell kernel to scan for partitions
  mkfs: fix final fsync and success/failure reports
  fsck: fix crash due to invalid sector size
  libexfat: fix endianness of GUID representation
  libexfat: ignore trailing garbage in exfat_count_used_clusters()
  libexfat.h: rearrange members of struct exfat_user_input
  treewide: use setlocale(LC_ALL, "")
  configure.ac: Fix build of dosattr programs when shell != bash

Fixes: https://github.com/LineageOS/issues/issues/364

+++ lib/exfat_dir.c
@@ -390,7 +390,7 @@ int exfat_de_iter_revert(struct exfat_de_iter *iter, int num)
 out:
 	iter->max_skip_dentries = 0;
 	iter->de_file_offset = file_offset;
-	iter->next_read_offset = (file_offset & ~(iter->read_size - 1)) + iter->read_size;
+	iter->next_read_offset = (file_offset & ~((off_t)iter->read_size - 1)) + iter->read_size;
 
 	return 0;
 }
@@ -763,8 +763,8 @@ int exfat_update_file_dentry_set(struct exfat *exfat,
 		}
 	}
 
-	dset[1].dentry.stream.valid_size = cpu_to_le64(ccount * exfat->clus_size);
-	dset[1].dentry.stream.size = cpu_to_le64(ccount * exfat->clus_size);
+	dset[1].dentry.stream.valid_size = cpu_to_le64((uint64_t)ccount * exfat->clus_size);
+	dset[1].dentry.stream.size = cpu_to_le64((uint64_t)ccount * exfat->clus_size);
 	if (start_clu)
 		dset[1].dentry.stream.start_clu = cpu_to_le32(start_clu);
